+++ eNMS/services/miscellaneous/rest_call.py
@@ -35,10 +35,11 @@ class RestCallService(Service):
     __mapper_args__ = {"polymorphic_identity": "rest_call_service"}
 
     def job(self, run, payload, device=None):
-        rest_url = run.sub(run.rest_url, locals())
+        local_variables = locals()
+        rest_url = run.sub(run.rest_url, local_variables)
         run.log("info", f"Sending REST Call to {rest_url}", device)
         kwargs = {
-            p: run.sub(getattr(self, p), locals())
+            p: run.sub(getattr(self, p), local_variables)
             for p in ("headers", "params", "timeout")
         }
         kwargs["verify"] = run.verify_ssl_certificate
@@ -47,7 +48,7 @@ class RestCallService(Service):
                 self.username, app.get_password(self.password)
             )
         if run.call_type in ("POST", "PUT", "PATCH"):
-            kwargs["json"] = run.sub(run.payload, locals())
+            kwargs["json"] = run.sub(run.payload, local_variables)
         call = getattr(app.request_session, run.call_type.lower())
         response = call(rest_url, **kwargs)
         if response.status_code not in range(200, 300):